Move create logic to UpdatePlanScreen, add order
